[4.2] fixed minor problems in OpenStep
seems to crash on closing windows though. I think NSMutableDictionary does not like you to remove keys that do not exist

+++ Soulver/SVRMathString.h
@@ -41,7 +41,7 @@
 
 @interface SVRMathString (NSObjectProtocol)
 -(BOOL)isEqual:(id)object;
--(unsigned long)hash;
+-(unsigned)hash;
 @end
 
 // MARK: Constants
